pyproject.toml
[tool.poetry]
name = "캡스톤"
version = "0.1.0"
description = ""
authors = ["SonJinHYo <wlsgy95@naver.com>"]
license = "\"MIT\""
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10.6"
djangorestframework = "^3.14.0"
wheel = "^0.40.0"
cmake = "^3.26.3"
gunicorn = "^20.1.0"
django = "^4.2"
uwsgi = "^2.0.21"
mysqlclient = "^2.1.1"
pymysql = "^1.0.3"
django-environ = "^0.10.0"
boto3 = "^1.26.126"
django-storages = "^1.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
